requirements.txt
Django==1.8.2
django-pagination==1.0.7
django-qiniu-storage==2.1.1
MySQL-python==1.2.3
pillow==2.9.0
python-memcached==1.54
qiniu==7.0.5
requests==2.7.0
six==1.9.0
wheel==0.24.0
